Learning/routing/multiple database/requirements.txt
langchain==0.3.23
langchain-community==0.3.21
langchain-text-splitters==0.3.8
langchain-google-genai==2.1.3
langchain-qdrant==0.2.0
qdrant-client==1.13.3
python-dotenv==1.1.0
openai==1.75.0
